+++ crates/renderer/src/lib.rs
@@ -49,8 +49,8 @@ pub use runtime::{
 };
 pub use types::{
     AdapterProfile, Antialiasing, ChannelBindings, ChannelSource, ChannelTextureKind,
-    ColorSpaceMode, GpuMemoryMode, GpuPowerPreference, RenderMode, RendererConfig, ShaderCompiler,
-    SurfaceAlpha, CUBEMAP_FACE_STEMS,
+    ColorSpaceMode, CrossfadeCurve, GpuMemoryMode, GpuPowerPreference, RenderMode, RendererConfig,
+    ShaderCompiler, SurfaceAlpha, CUBEMAP_FACE_STEMS,
 };
 pub use wallpaper::{
     OutputId, SurfaceId, SurfaceInfo, SurfaceSelector, SwapRequest, WallpaperRuntime,
